// src/frontend/ui/generated/ExtList.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExtMachine } from "./ExtMachine";
import type { ExtRow } from "./ExtRow";

/**
 * `ext-list` 的成品。
 */
export type ExtList = { machines: Array<ExtMachine>, rows: Array<ExtRow>, problems: Array<string>, };